Initial API commit

+++ src/core/application/usecases/user/deleteActivity.ts
@@ -1,5 +1,5 @@
-import { UserActivity } from "../../../domain/user/userActivity.ts";
-import type IUserRepository from "../../../application/ports/userRepository.ts";
+import { UserActivity } from "../../../user/userActivity.ts";
+import type IUserRepository from "../../repositories/userRepository.ts";
 
 export default class DeleteUserActivity{ 
     private repository: IUserRepository;
@@ -13,7 +13,12 @@ export default class DeleteUserActivity{
     }
 
     async execute(): Promise<void>{
-        return await this.repository.deleteUserActivity(this.userID, this.activityID);
-            
+        const foundUser = await this.repository.findUserByID(this.userID);
+        if(!foundUser){ throw new Error('User not found'); }
+
+        const foundActivity = await this.repository.findUserActivityByID(this.userID, this.activityID);
+        if(!foundActivity){ throw new Error('Activity not found'); }
+
+        await this.repository.deleteUserActivity(this.userID, this.activityID);
     }
 }
